+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   - na_ontap_ldap_client - Added options `group_dn` and `user_dn` in REST.
   - na_ontap_lun - updated documentation for unsupported REST parameters.
   - na_ontap_snapmirror - new options `time_out`, `wait_for_completion` added in REST.
+  - na_ontap_volume_efficiency- Added `wait_for_completion` and `time_out` parameters to fix time out errors for long running operations.
 
 ### Bug Fixes
   - na_ontap_vserver_audit - updated documentation for `log`.